코드 리팩토링 , api 호출 시 스켈레톤 로딩 처리
