+++ lib/constant.dart
@@ -59,3 +59,36 @@ const kCardContentTextStyle = TextStyle(
   fontWeight: FontWeight.w400,
   color: Colors.black
 );
+
+const buyOrNotCountTextStyle = TextStyle(
+  fontWeight: FontWeight.w500,
+  fontSize: 15,
+  height: 18 / 15,
+  color: Color(0xFF4F4F4F),
+);
+
+const dateStyle = TextStyle(
+  fontSize: 11,
+  height: 13 / 11,
+  color: Color(0XFFBDBDBD),
+);
+
+const stockDateStyle = TextStyle(
+  fontSize: 11,
+  height: 13 / 11,
+  color: Color(0xFF828282),
+);
+
+const stockTitleStyle =
+TextStyle(fontWeight: FontWeight.bold, fontSize: 22, height: 32 / 22);
+
+const stockBodyStyle = TextStyle(
+  fontSize: 14,
+  height: 20 / 14,
+);
+
+const stockBodyNumberStyle = TextStyle(
+  fontWeight: FontWeight.w500,
+  fontSize: 21,
+  height: 25 / 21,
+);
