+++ lib/screens/buy_or_not/buy_or_not_tab_screen.dart
@@ -1,17 +1,7 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:should_have_bought_app/constant.dart';
-import 'package:should_have_bought_app/models/buy_or_not/buy_or_not_stock.dart';
-import 'package:should_have_bought_app/models/buy_or_not/stock_hist.dart';
 import 'package:should_have_bought_app/models/calculator/company.dart';
-import 'package:should_have_bought_app/models/drip_room/evaluation_item.dart';
-import 'package:should_have_bought_app/providers/buy_or_not/buy_or_not_provider.dart';
-import 'package:should_have_bought_app/providers/calculator/calculator_widget_provider.dart';
-import 'package:should_have_bought_app/screens/tab_screen.dart';
-import 'package:should_have_bought_app/utils.dart';
-import 'package:should_have_bought_app/widgets/chart/buy_or_not_chart_widget.dart';
-import 'package:should_have_bought_app/widgets/login/login_handler.dart';
+import 'package:should_have_bought_app/widgets/buy_or_not/buy_or_not_select_widget.dart';
+import 'package:should_have_bought_app/widgets/buy_or_not/how_to_bought_then_widget.dart';
 
 class BuyorNotTabScreen extends StatelessWidget {
   final Company company;
@@ -34,333 +24,3 @@ class BuyorNotTabScreen extends StatelessWidget {
     );
   }
 }
-
-class HowToBoughtThenWidget extends StatelessWidget {
-  final Company company;
-
-  HowToBoughtThenWidget(this.company);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: MediaQuery.of(context).size.height * 0.7,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(14),
-        color: Colors.white,
-      ),
-      child: Consumer<BuyOrNotProvider>(
-          builder: (context, buyOrNotProvider, child) {
-        final StockHist stockHist = buyOrNotProvider.stockHist;
-        final EvaluationItem evaluationItem = buyOrNotProvider.evaluationItem;
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Stack(
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(top: 60),
-                  child: Container(
-                      height: MediaQuery.of(context).size.height * 0.35,
-                      child: BuyOrNotChartWidget(company.code)),
-                ),
-                Padding(
-                  padding: const EdgeInsets.all(29),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        '${commonDateFormat(evaluationItem.createdDate ?? '')}',
-                        style: dateStyle,
-                      ),
-                      Text(company.company ?? '',
-                          style: stockTitleStyle),
-                      Text(
-                          '${numberWithComma(stockHist.price?.toString() ?? '0')}원',
-                          style: stockTitleStyle),
-                      Text(
-                          "+2,500 (${stockHist.changeInPercent?.toString() ?? '-'}%)",
-                          style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                              fontSize: 14,
-                              height: 16 / 14,
-                              color: Color(0xFF4990FF))),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: [
-                Column(
-                  children: [
-                    Row(
-                      children: [
-                        Text('최근 10년간', style: stockBodyStyle),
-                        Text(
-                          ' 최고가',
-                          style: TextStyle(
-                            fontSize: 14,
-                            height: 20 / 14,
-                            color: likeColor,
-                          ),
-                        )
-                      ],
-                    ),
-                    Text(
-                        '${numberWithComma(stockHist.maxQuote?.high?.toString() ?? '0')}원',
-                        style: stockBodyNumberStyle),
-                    Text(
-                        '${commonDayDateFormat(stockHist.maxQuote?.date ?? '')} 종가 기준',
-                        style: stockDateStyle),
-                  ],
-                ),
-                Column(
-                  children: [
-                    Row(
-                      children: [
-                        Text('최근 10년간', style: stockBodyStyle),
-                        Text(
-                          ' 최저가',
-                          style: TextStyle(
-                            fontSize: 14,
-                            height: 20 / 14,
-                            color: unlikeColor,
-                          ),
-                        )
-                      ],
-                    ),
-                    Text(
-                        '${numberWithComma(stockHist.minQuote?.low?.toString() ?? '0')}원',
-                        style: stockBodyNumberStyle),
-                    Text(
-                        '${commonDayDateFormat(stockHist.minQuote?.date ?? '')} 종가 기준',
-                        style: stockDateStyle),
-                  ],
-                ),
-              ],
-            ),
-            SizedBox(height: 20),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                      primary: mainColor,
-                      onPrimary: mainColor,
-                      elevation: 0,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(
-                            16.0), //side: BorderSide(color: Colors.red)
-                      )),
-                  onPressed: () {
-                    print(company);
-                    Provider.of<CalculatorWidgetProvider>(context, listen: false)
-                        .setCompanyAndDateValue(Company(company: company.company, code: company.code), 'YEAR10');
-                    Navigator.pop(context);
-                    Navigator.pushReplacement(context,  MaterialPageRoute(builder: (context) => TabScreen(selectIndex: 0)));
-                  },
-                  child: SizedBox(
-                    height: 50,
-                    width: MediaQuery.of(context).size.width,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Text(
-                          '그때 샀으면 지금 얼마게?',
-                          style: TextStyle(
-                              fontWeight: FontWeight.w500,
-                              fontSize: 15,
-                              color: Colors.white),
-                        ),
-                      ],
-                    ),
-                  )),
-            )
-          ],
-        );
-      }),
-    );
-  }
-}
-
-class BuyorNotSelectWidget extends StatefulWidget {
-  final String stockCode;
-
-  BuyorNotSelectWidget(this.stockCode);
-
-  _CreateBuyorNotSelectWidgetState createState() =>
-      _CreateBuyorNotSelectWidgetState();
-}
-
-class _CreateBuyorNotSelectWidgetState extends State<BuyorNotSelectWidget> {
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-
-  @override
-  void initState() {
-    Provider.of<BuyOrNotProvider>(context, listen: false)
-        .getBuyOrNotStock(widget.stockCode);
-  }
-
-  void setEvaluation(String stockCode, String buySell) async {
-    await Provider.of<BuyOrNotProvider>(context, listen: false)
-        .setBuyOrNotStock(stockCode, buySell);
-  }
-
-  String parseWasEvaluation(BuyOrNotStock buyOrNotStock, String buySell) {
-    return buyOrNotStock.userBuySell == buySell ? 'NULL' : buySell;
-  }
-
-  void actionEvaluation(
-      BuildContext context, BuyOrNotStock buyOrNotStock, String buySell) async{
-    if (isNotLogin(_auth.currentUser)) {
-      LoginHandler(context, buyOrNotStockCode:buyOrNotStock.code);
-      return;
-    }
-    String _buySell = parseWasEvaluation(buyOrNotStock, buySell);
-    setEvaluation(buyOrNotStock.code, _buySell);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Consumer<BuyOrNotProvider>(
-        builder: (context, buyOrNotProvider, child) {
-      BuyOrNotStock buyOrNotStock = buyOrNotProvider.buyOrNotStock;
-      return Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(14),
-          color: Colors.white,
-        ),
-        height: 82,
-        child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-          InkWell(
-            onTap: () {
-              print('살래');
-              actionEvaluation(context, buyOrNotStock, 'BUY');
-            },
-            child: Wrap(
-              children: [
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    SizedBox(
-                      width: 50,
-                      height: 50,
-                      child: ClipOval(
-                        child: Container(
-                          color: buyOrNotStock.userBuySell == "BUY"
-                              ? likeColor
-                              : disableColor,
-                          padding: EdgeInsets.only(bottom: 5),
-                          child: Image(image: AssetImage('assets/icons/ico_like_big.png')),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-                SizedBox(width: 10),
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Text(
-                      '살래?',
-                      style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 14,
-                        height: 20 / 14,
-                        color: likeColor,
-                      ),
-                    ),
-                    Text("${buyOrNotStock?.buyCount ?? 0}",
-                        style: buyOrNotCountTextStyle),
-                  ],
-                ),
-              ],
-            ),
-          ),
-          SizedBox(width: 50),
-          InkWell(
-            onTap: () {
-              print('말래');
-              actionEvaluation(context, buyOrNotStock, 'SELL');
-            },
-            child: Wrap(
-              children: [
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    SizedBox(
-                      width: 50,
-                      height: 50,
-                      child: ClipOval(
-                        child: Container(
-                          color: buyOrNotStock.userBuySell == "SELL"
-                              ? nagativeColor
-                              : disableColor,
-                          child: Image(
-                              image: AssetImage(
-                                  'assets/icons/ico_unlike_big.png')),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-                SizedBox(width: 10),
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Text(
-                      '말래?',
-                      style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 14,
-                        height: 20 / 14,
-                        color: unlikeColor,
-                      ),
-                    ),
-                    Text("${buyOrNotStock?.sellCount ?? 0}",
-                        style: buyOrNotCountTextStyle),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ]),
-      );
-    });
-  }
-}
-
-const buyOrNotCountTextStyle = TextStyle(
-  fontWeight: FontWeight.w500,
-  fontSize: 15,
-  height: 18 / 15,
-  color: Color(0xFF4F4F4F),
-);
-
-const dateStyle = TextStyle(
-  fontSize: 11,
-  height: 13 / 11,
-  color: Color(0XFFBDBDBD),
-);
-
-const stockDateStyle = TextStyle(
-  fontSize: 11,
-  height: 13 / 11,
-  color: Color(0xFF828282),
-);
-
-const stockTitleStyle =
-    TextStyle(fontWeight: FontWeight.bold, fontSize: 22, height: 32 / 22);
-
-const stockBodyStyle = TextStyle(
-  fontSize: 14,
-  height: 20 / 14,
-);
-
-const stockBodyNumberStyle = TextStyle(
-  fontWeight: FontWeight.w500,
-  fontSize: 21,
-  height: 25 / 21,
-);
