+++ lib/screens/drip_room/drip_room_screen.dart
@@ -1,12 +1,14 @@
 import 'package:auto_size_text/auto_size_text.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:provider/provider.dart';
 import 'package:should_have_bought_app/constant.dart';
 import 'package:should_have_bought_app/models/calculator/company.dart';
 import 'package:should_have_bought_app/models/drip_room/evaluation_item.dart';
 import 'package:should_have_bought_app/providers/drip_room/drip_room_provider.dart';
+import 'package:should_have_bought_app/screens.dart';
+import 'package:should_have_bought_app/screens/util/skeleton_widget.dart';
 import 'package:should_have_bought_app/widgets/appbar/drip_room_appbar.dart';
 import 'package:should_have_bought_app/screens/stock/stock_detail_screen.dart';
 import 'package:should_have_bought_app/widgets/background/flat_background_frame.dart';
@@ -15,66 +17,82 @@ class DripRoomScreen extends StatefulWidget {
   @override
   _DripRoomScreenState createState() => _DripRoomScreenState();
 }
-class _DripRoomScreenState extends State<DripRoomScreen> {
 
+class _DripRoomScreenState extends State<DripRoomScreen> {
   @override
-  void didChangeDependencies() {
+  void didChangeDependencies() async{
     // TODO: implement didChangeDependencies
     super.didChangeDependencies();
-    Map<String, dynamic> parmeters = {'order':'LATELY','pageNo':'1', 'pageSize':'100'};
-    bool isEmpty = Provider.of<DripRoomProvider>(context, listen:false).evaluationItemList.length == 0 ? true : false;
-    if(isEmpty) {
+    Map<String, dynamic> parmeters = {
+      'order': 'LATELY',
+      'pageNo': '1',
+      'pageSize': '100'
+    };
+    bool isEmpty = Provider.of<DripRoomProvider>(context, listen: false)
+                .evaluationItemList
+                .length ==
+            0
+        ? true
+        : false;
+    if (isEmpty) {
+      await EasyLoading.show(
+        status: 'loading...',
+        maskType: EasyLoadingMaskType.none,
+      );
       Provider.of<DripRoomProvider>(context, listen: false)
-          .getEvaluationList(parmeters);
+          .getEvaluationList(parmeters).then((value) => EasyLoading.dismiss());
     }
   }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Color(0xFFF6F6F6),
       body: Consumer<DripRoomProvider>(
-        builder: (context, dripRoomProvider, child) {
-          List evaluationItemList = dripRoomProvider.evaluationItemList;
-          return Container(
-            margin:EdgeInsets.only(bottom: 100),
-            child: ListView.builder(
-                itemCount: evaluationItemList.length,
-                itemBuilder: (BuildContext context, int index) {
-                  if(index == 0) {
-                    return Column(
-                      children: [
-                        FlatBackgroundFrame(
-                            child: DripRoomWidget()
-                        ),
-                        SizedBox(height: 37),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 16),
-                          child: Column(
-                            children: [
-                              Row(
-                                mainAxisAlignment: MainAxisAlignment.end,
-                                children: [
-                                  Text('최신순'),
-                                  Icon(Icons.keyboard_arrow_down_outlined)
-                                ],
-                              ),
-                              SizedBox(height: 11),
-                              DripCardWidget(evaluationItemList[index])
-                            ],
-                          ),
-                        )
-                      ],
-                    );
-                  }
-                  return Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 16),
-                    child: DripCardWidget(evaluationItemList[index]),
+          child: emptyDripRoomScreen(context),
+          builder: (context, dripRoomProvider, child) {
+            List evaluationItemList = dripRoomProvider.evaluationItemList;
+            return dripRoomProvider.evaluationItemList.length == 0
+                ? child
+                : Container(
+                    margin: EdgeInsets.only(bottom: 100),
+                    child: ListView.builder(
+                        itemCount: evaluationItemList.length,
+                        itemBuilder: (BuildContext context, int index) {
+                          if (index == 0) {
+                            return Column(
+                              children: [
+                                FlatBackgroundFrame(child: DripRoomWidget()),
+                                SizedBox(height: 37),
+                                Padding(
+                                  padding: const EdgeInsets.symmetric(
+                                      horizontal: 16),
+                                  child: Column(
+                                    children: [
+                                      Row(
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.end,
+                                        children: [
+                                          Text('최신순'),
+                                          Icon(Icons
+                                              .keyboard_arrow_down_outlined)
+                                        ],
+                                      ),
+                                      SizedBox(height: 11),
+                                      DripCardWidget(evaluationItemList[index])
+                                    ],
+                                  ),
+                                )
+                              ],
+                            );
+                          }
+                          return Container(
+                            padding: const EdgeInsets.symmetric(horizontal: 16),
+                            child: DripCardWidget(evaluationItemList[index]),
+                          );
+                        }),
                   );
-              }
-            ),
-          );
-        }
-      ),
+          }),
     );
   }
 }
@@ -82,7 +100,6 @@ class _DripRoomScreenState extends State<DripRoomScreen> {
 class DripRoomWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 16),
       child: Column(
@@ -111,20 +128,27 @@ class DripRoomWidget extends StatelessWidget {
 
 class DripCardWidget extends StatelessWidget {
   final EvaluationItem evaluationItem;
+
   DripCardWidget(this.evaluationItem);
 
   @override
   Widget build(BuildContext context) {
     return InkWell(
       onTap: () {
-        Navigator.push(context, MaterialPageRoute(builder: (context) => StockDetailScreen(
-        Company(company:evaluationItem.company, code: evaluationItem.code),1)
-        ));
+        Navigator.push(
+            context,
+            MaterialPageRoute(
+                builder: (context) => StockDetailScreen(
+                    Company(
+                        company: evaluationItem.company ?? '',
+                        code: evaluationItem.code ?? ''),
+                    1)));
       },
       child: Container(
           width: MediaQuery.of(context).size.width,
-          padding: const EdgeInsets.only(left: 20, right: 20, top: 15, bottom: 26),
-          margin : const EdgeInsets.only(bottom: 15),
+          padding:
+              const EdgeInsets.only(left: 20, right: 20, top: 15, bottom: 26),
+          margin: const EdgeInsets.only(bottom: 15),
           decoration: BoxDecoration(
             color: Colors.white,
             borderRadius: BorderRadius.circular(25),
@@ -132,21 +156,25 @@ class DripCardWidget extends StatelessWidget {
           child: Column(
             children: [
               Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 13.0,),
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 13.0,
+                ),
                 child: Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
                     Text(
-                      evaluationItem.company,
-                      style: TextStyle(fontSize: 20, fontWeight: FontWeight.w500),
+                      evaluationItem.company ?? '',
+                      style:
+                          TextStyle(fontSize: 20, fontWeight: FontWeight.w500),
                     ),
                     Column(
                       crossAxisAlignment: CrossAxisAlignment.center,
                       children: [
                         Padding(
                           padding: const EdgeInsets.only(bottom: 2, right: 4),
-                          child: Image(image: AssetImage('assets/icons/ico_like.png')),
+                          child: Image(
+                              image: AssetImage('assets/icons/ico_like.png')),
                         ),
                         Text(
                           evaluationItem.likeCount.toString(),
@@ -167,8 +195,8 @@ class DripCardWidget extends StatelessWidget {
               Padding(
                 padding: const EdgeInsets.only(left: 13.0),
                 child: ProsAndConsWidget(
-                  pros: evaluationItem.pros,
-                  cons: evaluationItem.cons,
+                  pros: evaluationItem?.pros ?? '',
+                  cons: evaluationItem?.cons ?? '',
                 ),
               ),
             ],
@@ -176,16 +204,23 @@ class DripCardWidget extends StatelessWidget {
     );
   }
 }
+
 class BestDripCardWidget extends StatefulWidget {
   @override
-  _CreateBestDripCardWidgetState createState() => _CreateBestDripCardWidgetState();
+  _CreateBestDripCardWidgetState createState() =>
+      _CreateBestDripCardWidgetState();
 }
+
 class _CreateBestDripCardWidgetState extends State<BestDripCardWidget> {
+  bool _isLoading = false;
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    Provider.of<DripRoomProvider>(context, listen: false).getTodayBest();
+    setState(() => _isLoading = true);
+    Provider.of<DripRoomProvider>(context, listen: false)
+        .getTodayBest()
+        .then((value) => setState(() => _isLoading = false));
   }
 
   @override
@@ -198,59 +233,68 @@ class _CreateBestDripCardWidgetState extends State<BestDripCardWidget> {
         borderRadius: BorderRadius.circular(12),
       ),
       child: Consumer<DripRoomProvider>(
-        builder: (context, dripRoomProvider, child) {
-          EvaluationItem evaluationItem = dripRoomProvider.todayBest;
-          return Column(
-            children: [
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 13.0,),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      evaluationItem?.company ?? '',
-                      style: TextStyle(fontSize: 20, fontWeight: FontWeight.w500),
-                    ),
-                    Column(
-                      crossAxisAlignment: CrossAxisAlignment.end,
-                      children: [
-                        Text(
-                          evaluationItem?.displayName ?? '',
+          builder: (context, dripRoomProvider, child) {
+        EvaluationItem evaluationItem = dripRoomProvider.todayBest;
+        return Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.symmetric(
+                horizontal: 13.0,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  _isLoading
+                      ? skeletonText(100, 25)
+                      : Text(
+                          evaluationItem?.company ?? '',
                           style: TextStyle(
-                              fontSize: 12,
-                              fontWeight: FontWeight.w500,
-                              height: 17 / 12),
+                              fontSize: 20, fontWeight: FontWeight.w500),
                         ),
-                        Text(
-                          evaluationItem?.createdDate ?? '',
-                          style: TextStyle(
-                              fontSize: 11,
-                              fontWeight: FontWeight.w300,
-                              color: Color(0xFF828282)),
-                        )
-                      ],
-                    ),
-                  ],
-                ),
+                  Column(
+                    crossAxisAlignment: CrossAxisAlignment.end,
+                    children: [
+                      _isLoading
+                          ? skeletonText(40, 15)
+                          : Text(
+                              evaluationItem?.displayName ?? '',
+                              style: TextStyle(
+                                  fontSize: 12,
+                                  fontWeight: FontWeight.w500,
+                                  height: 17 / 12),
+                            ),
+                      _isLoading
+                          ? skeletonText(100, 15)
+                          : Text(
+                              evaluationItem?.createdDate ?? '',
+                              style: TextStyle(
+                                  fontSize: 11,
+                                  fontWeight: FontWeight.w300,
+                                  color: Color(0xFF828282)),
+                            )
+                    ],
+                  ),
+                ],
               ),
-              SizedBox(height: 9),
-              Divider(
-                color: Color(0xFF8E8E8E),
-                thickness: 0.6,
+            ),
+            SizedBox(height: 9),
+            Divider(
+              color: Color(0xFF8E8E8E),
+              thickness: 0.6,
+            ),
+            SizedBox(height: 9),
+            Padding(
+              padding: const EdgeInsets.only(left: 13.0),
+              child: ProsAndConsWidget(
+                pros: evaluationItem?.pros ?? '',
+                cons: evaluationItem?.cons ?? '',
+                isLoading: _isLoading,
               ),
-              SizedBox(height: 9),
-              Padding(
-                padding: const EdgeInsets.only(left: 13.0),
-                child: ProsAndConsWidget(
-                  pros: evaluationItem?.pros ?? '',
-                  cons: evaluationItem?.cons ?? '',
-                ),
-              )
-            ],
-          );
-        }
-      ),
+            )
+          ],
+        );
+      }),
     );
   }
 }
@@ -258,7 +302,10 @@ class _CreateBestDripCardWidgetState extends State<BestDripCardWidget> {
 class ProsAndConsWidget extends StatelessWidget {
   final String pros;
   final String cons;
-  ProsAndConsWidget({@required this.pros, @required this.cons});
+  final bool isLoading;
+
+  ProsAndConsWidget(
+      {@required this.pros, @required this.cons, @required this.isLoading});
 
   @override
   Widget build(BuildContext context) {
@@ -283,16 +330,20 @@ class ProsAndConsWidget extends StatelessWidget {
               ),
             ),
             SizedBox(width: 11),
-            Expanded(
-              child: AutoSizeText(
-                pros ?? '',
-                style: TextStyle(
-                fontWeight: FontWeight.w400, fontSize: 14, height: 20 / 14),
-                presetFontSizes: [14, 13],
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ),
+            (isLoading != null && isLoading)
+                ? skeletonText(80, 15)
+                : Expanded(
+                    child: AutoSizeText(
+                      pros ?? '',
+                      style: TextStyle(
+                          fontWeight: FontWeight.w400,
+                          fontSize: 14,
+                          height: 20 / 14),
+                      presetFontSizes: [14, 13],
+                      maxLines: 1,
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                  ),
           ],
         ),
         SizedBox(
@@ -317,19 +368,67 @@ class ProsAndConsWidget extends StatelessWidget {
               ),
             ),
             SizedBox(width: 11),
-            Expanded(
-              child: AutoSizeText(
-                cons ?? '',
-                style: TextStyle(
-                    fontWeight: FontWeight.w400, fontSize: 14, height: 20 / 14),
-                presetFontSizes: [14, 13],
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ),
+            (isLoading != null && isLoading)
+                ? skeletonText(80, 15)
+                : Expanded(
+                    child: AutoSizeText(
+                      cons ?? '',
+                      style: TextStyle(
+                          fontWeight: FontWeight.w400,
+                          fontSize: 14,
+                          height: 20 / 14),
+                      presetFontSizes: [14, 13],
+                      maxLines: 1,
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                  ),
           ],
         )
       ],
     );
   }
+}
+
+
+Widget emptyDripRoomScreen(BuildContext context) {
+  return Column(
+    children: [
+      FlatBackgroundFrame(
+          child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 16),
+              child: Column(
+                children: [
+                  DripRoomAppBar(context),
+                  SizedBox(height: 33),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.start,
+                    children: [
+                      Text(
+                        "Today's BEST",
+                        style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 15,
+                            fontWeight: FontWeight.bold),
+                      ),
+                    ],
+                  ),
+                  SizedBox(height: 9),
+                ],
+              ))),
+      SizedBox(height: 37),
+      Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 16),
+        child: Column(
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [Text('최신순'), Icon(Icons.keyboard_arrow_down_outlined)],
+            ),
+            SizedBox(height: 11),
+            DripCardWidget(EvaluationItem())
+          ],
+        ),
+      )
+    ],
+  );
 }
