+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   flutter_easyloading: ^3.0.0
   carousel_slider: ^3.0.0
   emoji_picker_flutter: ^1.0.4
+  skeleton_text: ^3.0.0
 
 dev_dependencies:
   flutter_test:
